# 2025-12-22-cqrs-demo/packages/backend/src/main/java/com/example/demo/presentation/TodoController.java
package com.example.demo.presentation;

import com.example.demo.application.command.TodoCommandService;
import com.example.demo.application.query.TodoQueryService;
import com.example.demo.application.query.TodoView;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.time.LocalDate;
import java.util.List;
import java.util.UUID;


@RestController
@RequestMapping("/api/todos")
public class TodoController {
    private final TodoCommandService commandService;
    private final TodoQueryService queryService;

    public TodoController(TodoCommandService commandService, TodoQueryService queryService) {
        this.commandService = commandService;
        this.queryService = queryService;
    }

    @PostMapping
    public ResponseEntity<CreateTodoResponse> createTodo(@RequestBody CreateTodoRequest request) {
        UUID id = commandService.createTodo(request.title(), request.description(), request.dueDate());
        return ResponseEntity.status(HttpStatus.CREATED).body(new CreateTodoResponse(id));
    }

    @GetMapping
    public ResponseEntity<List<TodoView>> getAllTodos() {
        List<TodoView> todos = queryService.getAllTodos();
        return ResponseEntity.ok(todos);
    }

    @GetMapping("/{id}")
    public ResponseEntity<TodoView> getTodo(@PathVariable UUID id) {
        TodoView todo = queryService.getTodo(id);
        return ResponseEntity.ok(todo);
    }

    @PutMapping("/{id}/complete")
    public ResponseEntity<Void> completeTodo(@PathVariable UUID id) {
        commandService.completeTodo(id);
        return ResponseEntity.ok().build();
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Void> deleteTodo(@PathVariable UUID id) {
        commandService.deleteTodo(id);
        return ResponseEntity.ok().build();
    }

    @PutMapping("/{id}/update")
    public ResponseEntity<Void> updateTodo(@PathVariable UUID id, @RequestBody UpdateTodoRequest request) {
        commandService.updateTodo(id, request.title(), request.description(), request.dueDate());
        return ResponseEntity.noContent().build();
    }

    public record CreateTodoRequest(String title, String description, LocalDate dueDate) {}
    public record UpdateTodoRequest(String title, String description, LocalDate dueDate) {}
    public record CreateTodoResponse(UUID id) {}
}