# 2025-12-22-cqrs-demo/packages/shared/openapi.yaml
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/todos/{id}/update":{"put":{"tags":["todo-controller"],"operationId":"updateTodo","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTodoRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/todos/{id}/complete":{"put":{"tags":["todo-controller"],"operationId":"completeTodo","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/todos":{"get":{"tags":["todo-controller"],"operationId":"getAllTodos","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TodoView"}}}}}}},"post":{"tags":["todo-controller"],"operationId":"createTodo","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTodoRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateTodoResponse"}}}}}}},"/api/todos/{id}":{"get":{"tags":["todo-controller"],"operationId":"getTodo","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TodoView"}}}}}},"delete":{"tags":["todo-controller"],"operationId":"deleteTodo","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"UpdateTodoRequest":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"dueDate":{"type":"string","format":"date"}}},"CreateTodoRequest":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"dueDate":{"type":"string","format":"date"}}},"CreateTodoResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"}}},"TodoView":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"title":{"type":"string"},"description":{"type":"string"},"completed":{"type":"boolean"},"dueDate":{"type":"string","format":"date"}}}}}}